pyproject.toml
[project]
name = "mimir_news"
version = "0.1.0"
description = "Evaluate news based on interest in prediction markets"
authors = [{ name = "Eskild Eriksen", email = "eskild.eriksen122@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "elevenlabs",
    "moviepy",
    "openai",
    "pandas",
    "prompt-poet",
    "pydantic",
    "pymanifold @ git+https://github.com/iameskild/pymanifold.git",
    "python-dotenv",
    "requests",
    "selenium",
]

[project.scripts]
mimir_news = "mimir_news.main:main"
# run_crew = "mimir_news.main:run"
# train_crew = "mimir_news.main:train"
# replay_crew = "mimir_news.main:replay"
# test_crew = "mimir_news.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
